Add jar generation and move runners into right packages
