+++ Introduction/AlmostDone/FinishTheGame/src/main/kotlin/almostdone/gif/GifHelper.kt
