+++ gifs/src/main/kotlin/org/jetbrains/kotlin/course/runners/almost/done/configs/ChoosePictureConfig.yaml
